style: move node manipulation into separate file
