+++ src/utils/zon/index.ts
@@ -1,4 +1,5 @@
 export * from "./types.ts";
+export * from "./node.ts";
 export * from "./path.ts";
 export * from "./stats.ts";
 export * from "./zon.ts";
